Move all data transformation into seperate files for Jordan tool
